backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.42.0
langchain>=0.3.14
langchain-anthropic>=0.3.4
langchain-core>=0.3.0
langgraph>=0.2.0
tavily-python>=0.5.0
httpx>=0.27.0
duckduckgo-search>=6.0.0
